feat: check produto and estoque

+++ src/data/services/deletar-produto-estoque.ts
@@ -1,16 +1,21 @@
 import { AdicionarProdutoEstoqueModel } from "../../domain/models/adicionar-produto-estoque"
 import { DeletarProdutoEstoqueUseCase } from "../../domain/useCases/deletar-produto-estoque"
+import { AMQPClient } from "../contracts/amqp-client"
+import { EstoqueRepository } from "../contracts/estoque-repository"
 import { ProdutoEstoqueRepository } from "../contracts/produtoEstoque-repository"
-import { Estoque } from "../entities/estoque"
-import { Produto } from "../entities/produto"
 
 export class DeletarProdutoEstoqueService implements DeletarProdutoEstoqueUseCase {
-    constructor (private readonly produtoEstoqueRepository: ProdutoEstoqueRepository) {}
+    constructor (private readonly amqpClient: AMQPClient, private readonly estoqueRepository: EstoqueRepository, private readonly produtoEstoqueRepository: ProdutoEstoqueRepository) {}
     async deletar (data: AdicionarProdutoEstoqueModel): Promise<void | Error> {
-        const produto = new Produto()
-        produto.id = data.idProduto
-        const estoque = new Estoque()
-        estoque.id = data.idEstoque
+        const produto = await this.amqpClient.send('buscar-produto', { id: data.idProduto })
+        if (produto && produto['error']) {
+            return new Error(produto['error'])
+        }
+
+        const estoque = await this.estoqueRepository.getById(data.idEstoque)
+        if (!estoque) {
+            return new Error('Esse estoque não foi encontrado!')
+        }
         await this.produtoEstoqueRepository.delete({
             produto,
             estoque
